feat: Persist URL tokens
The application may need to restart at times, and then we loose all the
data. To prevent that, we need to persist the data we've created.

Using sqlite3 is the most handy solution because it is part of the
Python distribution. Our system is super simple for anything bigger than
sqlite.

Adding persistent via transactional database has the benefit that
multiple app instances can communicate too.